cleanup: refactor code to make it more clear
