+++ kubespider/app.py
@@ -1,8 +1,8 @@
-from core import kubespider
+from core import runner
 
 
 def main():
-    kubespider.run_with_config_handler()
+    runner.run_with_config_handler()
 
 if __name__ == "__main__":
     main()
